+++ src/main/java/org/folio/circulation/resources/ExpiredSessionProcessingResource.java
@@ -1,8 +1,10 @@
 package org.folio.circulation.resources;
 
+import java.util.List;
 import static org.folio.circulation.support.Result.failed;
 
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
 import org.folio.circulation.domain.ConfigurationRepository;
@@ -38,7 +40,6 @@ public class ExpiredSessionProcessingResource extends Resource {
   }
 
   private void process(RoutingContext routingContext) {
-
     final WebContext context = new WebContext(routingContext);
     final Clients clients = Clients.create(context, client);
 
@@ -50,10 +51,12 @@ public class ExpiredSessionProcessingResource extends Resource {
 
     configurationRepository.lookupSessionTimeout()
       .thenCompose(r -> r.after(this::defineExpiredTime))
-      .thenCompose(r -> patronExpiredSessionRepository.findPatronExpiredSessions(PatronActionType.ALL, r.value().toString()))
-      .thenCompose(r -> r.after(expiredSession -> attemptEndSession(patronSessionService, expiredSession))
-        .thenApply(this::createWritableResult)
-        .thenAccept(result -> result.writeTo(routingContext.response())));
+      .thenCompose(r -> patronExpiredSessionRepository.findPatronExpiredSessions(
+        PatronActionType.ALL, r.value().toString()))
+      .thenCompose(r -> r.after(expiredSessions -> attemptEndSession(
+        patronSessionService, expiredSessions)))
+      .thenApply(this::createWritableResult)
+      .thenAccept(result -> result.writeTo(routingContext.response()));
   }
 
   private CompletableFuture<Result<DateTime>> defineExpiredTime(Integer timeout) {
@@ -62,12 +65,18 @@ public class ExpiredSessionProcessingResource extends Resource {
     return CompletableFuture.completedFuture(dateTimeResult);
   }
 
-  private CompletableFuture<Result<Void>> attemptEndSession(PatronActionSessionService patronSessionService,
-                                                            ExpiredSession expiredSession) {
-    if (expiredSession == null || StringUtils.isBlank(expiredSession.getPatronId())) {
+  private CompletableFuture<Result<Void>> attemptEndSession(
+    PatronActionSessionService patronSessionService, List<ExpiredSession> expiredSessions) {
+
+    List<ExpiredSession> existingExpiredSessions = expiredSessions.stream()
+      .filter(session -> StringUtils.isNotBlank(session.getPatronId()))
+      .collect(Collectors.toList());
+
+    if (existingExpiredSessions.isEmpty()) {
       return CompletableFuture.completedFuture(Result.succeeded(null));
     }
-    return patronSessionService.endSession(expiredSession.getPatronId(), expiredSession.getActionType());
+
+    return patronSessionService.endSession(expiredSessions);
   }
 
   private ResponseWritableResult<Void> createWritableResult(Result<?> result) {